changed scroll tracking to homepage
