+++ client/src/pages/HomePage.js
@@ -1,4 +1,7 @@
-import React from 'react';
+import React, {useState} from 'react';
+
+// API Calls
+import { getLimitedPosts } from './../api/handlePost';
 
 // Components
 import About from './../components/About';
@@ -7,9 +10,31 @@ import LoginModule from './../components/LoginModule';
 
 // Redux
 import { connect } from "react-redux";
+import { addNewPostToAllPostStore } from './../redux/actions/postActions';
+
+// CSS
+import styles from './../sass/pages/HomePage.module.scss';
 
 
 const HomePage = (props) => {
+
+  const [pageNumber, setPageNumber] = useState(1);
+
+  const onScroll = async (event) => {
+    let element = event.target
+    if (element.scrollHeight - element.scrollTop === element.clientHeight) {
+      // const id = props.allPosts[props.allPosts.length -1]._id;
+
+      setPageNumber(pageNumber + 1)
+      console.log(pageNumber)
+      await getLimitedPosts(pageNumber).then(async (response) => {
+        const newPosts = response.data;
+        props.addNewPostToAllPostStore(newPosts)
+      }).catch((err) => {
+        console.log(err)
+      })
+    }
+  } 
 
   const handlePopUp = (displayPopUpLogin, displayPopUpSignUp) => {
     if (displayPopUpSignUp === true) {
@@ -29,17 +54,23 @@ const HomePage = (props) => {
     } else {
       return (
         <div>
-          <About /> 
+          <About/> 
         </div>
       )
     }
   };
 
   return (
-    <div>
+    <div className={styles.mainContainer} onScroll={onScroll}>
       {handlePopUp(props.displayPopUpLogin, props.displayPopUpSignUp)} 
     </div>
   )
+};
+
+const mapDispatchToProps = (dispatch) => {
+  return {
+    addNewPostToAllPostStore: (newPost) => dispatch(addNewPostToAllPostStore(newPost)) 
+  }
 };
 
 // Connects to redux store so don't need to import.
@@ -50,5 +81,5 @@ function mapStateToProps(state) {
   };
 };
 
-export default connect(mapStateToProps)(HomePage);
+export default connect(mapStateToProps, mapDispatchToProps)(HomePage);
