+++ client/src/sass/pages/HomePage.module.scss
@@ -1,0 +1,4 @@
+.mainContainer {
+  height: 100vh;
+  overflow-y: scroll;
+}
