--- client/src/sass/pages/LandingPage.module.scss
@@ -1,1 +1,0 @@
-@import './../main.scss';
